Improve compatibility with GeckoDriver when using Selenium 3
moveto is not part of the W3C Webdriver spec, and so GeckoDriver does not
implement it.
Our click() implementation uses it to ensure that the element is scrolled into
the viewport before clicking on it, due to Selenium requirements.
Instead of forbidding the click entirely when using Geckodriver, this now
attempts the click directly when moving is unsupported, hoping that the
driver has not implemented the restriction when they don't have a way to
force the move yet.